+++ src/main/java/kg/attractor/job_search/service/UserService.java
@@ -12,5 +12,5 @@ public interface UserService {
 
     List<UserDto> getAllUsers();
 
-    UserDto getUserById(Long id);
+    UserDto getUserById(Integer id);
 }
